sqs working locally and bounty model is closer to what we want
