+++ contract_subscriber/v2/package.json
@@ -26,7 +26,7 @@
     "express": "^4.16.2",
     "redis": "^2.8.0",
     "rollbar": "^2.3.9",
-    "web3": "1.0.0-beta.34"
+    "web3": "^1.0.0-beta.26"
   },
   "devDependencies": {
     "chai": "^4.1.2",
