+++ contract_subscriber/v2/src/rollbar.js
@@ -1,6 +1,6 @@
 const Rollbar = require('rollbar');
 
-const rollbarToken = process.env['rollbar_token'] || "c0f6fabe29704bf38bccd403c8bfbe19";
+const rollbarToken = process.env['rollbar_token'];
 const rollbar = new Rollbar({
 	enabled: rollbarToken ? true : false,
     payload: {
